Add drinks to auto-suggest
Users can now search for drinks by name and select them directly from the Imbible.
Doing this with react router requires the use of the useHistory hook. I went ahead and dropped the use of `connect` in the coomponent and switched redux over to hooks as well.
I also noticed that the logic for rendering a drink was bugged - somewhere along the way we broke the assumption that drinks are sorted in id order.